docs: add Fermi edge correction workflows
