+++ skills/arpes-analysis/references/docs-links.md
@@ -16,9 +16,10 @@ available locally.
 - [Data loading and saving](https://erlabpy.readthedocs.io/en/stable/how-to/python/loading-and-saving.html)
 - [Data inspection and selection](https://erlabpy.readthedocs.io/en/stable/how-to/python/inspection-and-selection.html)
 - [Momentum conversion](https://erlabpy.readthedocs.io/en/stable/how-to/python/momentum-conversion.html)
+- [Fermi edge correction](https://erlabpy.readthedocs.io/en/stable/how-to/python/fermi-edge-correction.html)
 - [Curve fitting](https://erlabpy.readthedocs.io/en/stable/how-to/python/curve-fitting.html)
 - [Plotting](https://erlabpy.readthedocs.io/en/stable/how-to/plotting/index.html)
-- [Fit a reference when one energy range does not suit every EDC](https://erlabpy.readthedocs.io/en/stable/how-to/python/curve-fitting.html#how-to-python-fit-fermi-edge-separate-ranges)
+- [Fit a reference when one energy range does not suit every EDC](https://erlabpy.readthedocs.io/en/stable/how-to/python/fermi-edge-correction.html#how-to-python-fit-fermi-edge-separate-ranges)
 - [Annotating core levels](https://erlabpy.readthedocs.io/en/stable/how-to/plotting/core-levels.html)
 - [Transformations and filtering](https://erlabpy.readthedocs.io/en/stable/how-to/python/transformations-and-filtering.html)
 - [ARPES data conventions](https://erlabpy.readthedocs.io/en/stable/explanation/data-conventions.html)
@@ -32,6 +33,7 @@ available locally.
 
 - [Python and GUI workflows](https://erlabpy.readthedocs.io/en/stable/explanation/python-and-gui-workflows.html)
 - [Derived results and workspaces](https://erlabpy.readthedocs.io/en/stable/how-to/gui/derived-results-and-workspaces.html)
+- [Fermi edge correction in ImageTool Manager](https://erlabpy.readthedocs.io/en/stable/how-to/gui/fermi-edge-correction.html)
 - [ImageTool](https://erlabpy.readthedocs.io/en/stable/reference/gui/imagetool.html)
 - [Manager extensions](https://erlabpy.readthedocs.io/en/stable/how-to/gui/extensions.html)
 - [Manager extension reference](https://erlabpy.readthedocs.io/en/stable/reference/gui/extensions.html)
